requirements.txt
Django
fastapi
uvicorn[standard]
sqlmodel
httpx
passlib[bcrypt]
PyJWT
aio-pika